Can now execute the bootloader for early NTSC games

+++ src/libazote/op/op_alu.c
@@ -20,7 +20,7 @@ AZOTE_PROTO_R(azOpSUB)
 {
     if (!rd) return;
 
-    uint32_t value = state->cpu.registers[rs] - state->cpu.registers[rt];
+    uint32_t value = (uint32_t)state->cpu.registers[rs] - (uint32_t)state->cpu.registers[rt];
     state->cpu.registers[rd] = (int32_t)value;
 }
 
@@ -28,7 +28,7 @@ AZOTE_PROTO_R(azOpSUBU)
 {
     if (!rd) return;
 
-    uint32_t value = state->cpu.registers[rs] - state->cpu.registers[rt];
+    uint32_t value = (uint32_t)state->cpu.registers[rs] - (uint32_t)state->cpu.registers[rt];
     state->cpu.registers[rd] = (int32_t)value;
 }
